Split plotting module in submodules
Fix #338.

* Rename OrbitPlotter to StaticOrbitPlotter
* Only expose OrbitPlotter2D/3D in poliastro.plotting
* Make plot_solar_system use OrbitPlotter2D/3D
* Split tests

Some tests failing.
