+++ src/poliastro/plotting/__init__.py
@@ -1,0 +1,1 @@
+from .core import OrbitPlotter2D, OrbitPlotter3D
