+++ src/poliastro/tests/tests_plotting/test_static.py
@@ -1,21 +1,20 @@
 import astropy.units as u
 import matplotlib.pyplot as plt
-import pytest
 
-from poliastro.bodies import Earth, Jupiter, Mars
+from poliastro.bodies import Earth, Mars
 from poliastro.examples import iss
-from poliastro.plotting import OrbitPlotter, plot_solar_system
+from poliastro.plotting.static import StaticOrbitPlotter
 from poliastro.twobody.orbit import Orbit
 
 
 def test_orbitplotter_has_axes():
     ax = "Unused axes"
-    op = OrbitPlotter(ax)
+    op = StaticOrbitPlotter(ax)
     assert op.ax is ax
 
 
 def test_set_frame():
-    op = OrbitPlotter()
+    op = StaticOrbitPlotter()
     p = [1, 0, 0] * u.one
     q = [0, 1, 0] * u.one
     w = [0, 0, 1] * u.one
@@ -26,7 +25,7 @@ def test_set_frame():
 
 def test_axes_labels_and_title():
     ax = plt.gca()
-    op = OrbitPlotter(ax)
+    op = StaticOrbitPlotter(ax)
     ss = iss
     op.plot(ss)
 
@@ -35,7 +34,7 @@ def test_axes_labels_and_title():
 
 
 def test_number_of_lines_for_osculating_orbit():
-    op1 = OrbitPlotter()
+    op1 = StaticOrbitPlotter()
     ss = iss
 
     l1 = op1.plot(ss)
@@ -44,7 +43,7 @@ def test_number_of_lines_for_osculating_orbit():
 
 
 def test_legend():
-    op = OrbitPlotter()
+    op = StaticOrbitPlotter()
     ss = iss
     op.plot(ss, label="ISS")
     legend = plt.gca().get_legend()
@@ -56,7 +55,7 @@ def test_legend():
 
 
 def test_color():
-    op = OrbitPlotter()
+    op = StaticOrbitPlotter()
     ss = iss
     c = "#FF0000"
     op.plot(ss, label="ISS", color=c)
@@ -67,14 +66,8 @@ def test_color():
         assert element.get_c() == c
 
 
-@pytest.mark.parametrize("outer,expected", [(True, 8), (False, 4)])
-def test_plot_solar_system(outer, expected):
-    assert len(plot_solar_system(outer).orbits) == expected
-    assert isinstance(plot_solar_system(), OrbitPlotter)
-
-
 def test_plot_trajectory_sets_label():
-    op = OrbitPlotter()
+    op = StaticOrbitPlotter()
     earth = Orbit.from_body_ephem(Earth)
     mars = Orbit.from_body_ephem(Mars)
     trajectory = earth.sample()
@@ -85,23 +78,7 @@ def test_plot_trajectory_sets_label():
 
 
 def test_dark_mode_plots_dark_plot():
-    op = OrbitPlotter(dark=True)
+    op = StaticOrbitPlotter(dark=True)
     assert op.ax.get_facecolor() == (0.0, 0.0, 0.0, 1.0)
-    op = OrbitPlotter()
+    op = StaticOrbitPlotter()
     assert op.ax.get_facecolor() == (1.0, 1.0, 1.0, 1)
-
-
-def test_redraw_makes_attractor_none():
-    op = plot_solar_system()
-    op._redraw()
-    assert op._attractor_radius is not None
-
-
-def test_set_frame_plots_same_colors():
-    op = plot_solar_system()
-    jupiter = Orbit.from_body_ephem(Jupiter)
-    op.plot(jupiter)
-    colors1 = [orb[2] for orb in op._orbits]
-    op.set_frame(*jupiter.pqw())
-    colors2 = [orb[2] for orb in op._orbits]
-    assert colors1 == colors2
